dzialajaca wersja dla user i transaction
