+++ src/main/resources/TransactionDatabase.txt
@@ -23,3 +23,5 @@
 299.0, USD, PLN, 4.2276347472, 1264.0627894128, 2020/05/15 20:30:13, 8971, a, a, a
 888.0, USD, PLN, 4.2276347472, 3754.1396555136, 2020/05/15 20:30:43, 3152, b, b, b
 34.0, EUR, USD, 1.0798, 36.7132, 2020/05/15 21:51:39, 3295, we, we, we
+12345.0, EUR, USD, 1.0798, 13330.131000000001, 2020/05/15 23:07:10, 4441, et, et, et
+21.0, EUR, PLN, 4.565, 95.86500000000001, 2020/05/15 23:08:16, 4441, et, et, et
