+++ src/main/resources/WalletDatabase.txt
@@ -1,20 +1,21 @@
 0.0, 0, adminLogin, adminName, adminSurname
 1.0, 1, admin, admin, admin
 10.0, 6326, b, b, b
-0.0, 1515, JaneczekKowalik, Jan, Kowal
-0.0, 2634, dasd, zxc, sdfgv
-0.0, 1724, vcvb, hgf, n
-0.0, 6236, ahkj, jbh, jn
-0.0, 1650, ds, vxbcv, bdfg
-0.0, 8825, a, a, a
-0.0, 2441, a, a, a
-0.0, 1251, b, b, b
-0.0, 8635, sad, fvcb, dfg
-0.0, 1736, r, r, r
-0.0, 8943, dfsd, xcv, sdfg
-0.0, 75, t, t, t
-0.0, 739, q, q, q
-0.0, 5679, dsf, cvx, dfg
-0.0, 8778, afsdfsd, cxv, dfg
-0.0, 8039, sad, c, xzcv
-0.0, 3295, we, we, we
+5.0, 1515, JaneczekKowalik, Jan, Kowal
+6.0, 2634, dasd, zxc, sdfgv
+7.0, 1724, vcvb, hgf, n
+8.0, 6236, ahkj, jbh, jn
+9.0, 1650, ds, vxbcv, bdfg
+10.0, 8825, a, a, a
+110.0, 2441, a, a, a
+120.0, 1251, b, b, b
+130.0, 8635, sad, fvcb, dfg
+140.0, 1736, r, r, r
+150.0, 8943, dfsd, xcv, sdfg
+160.0, 75, t, t, t
+170.0, 739, q, q, q
+180.0, 5679, dsf, cvx, dfg
+190.0, 8778, afsdfsd, cxv, dfg
+20.0, 8039, sad, c, xzcv
+210.0, 3295, we, we, we
+0.0, 4441, et, et, et
